# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Surveyvor App - Technical Manual",
        "version": "1.0"
    },
    "paths": {
...
        "/api/nextQuestion/all/{survey_id}": {
            "get": {
                "tags": [
                    "Next Question"
                ],
                "summary": "Get all the Next Question from a survey",
                "description": "Returns a list of Next Questions",
                "operationId": "getNextQuestionsFromSurvey",
                "parameters": [
                    {
                        "name": "survey_id",
                        "in": "path",
                        "description": "Search by survey_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "definedAnswer": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "next_question_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "defined_answer_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "question_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/nextQuestion/save": {
            "post": {
                "tags": [
                    "Next Question"
                ],
                "summary": "Create a new Next Question",
                "description": "Returns the generated Next Question",
                "operationId": "saveNextQuestion",
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "defined_answer_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "question_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "question": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "next_question_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "defined_answer_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "question_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/nextQuestion/update/{next_question_id}": {
            "post": {
                "tags": [
                    "Next Question"
                ],
                "summary": "Update a Next Question",
                "operationId": "updateNextQuestion",
                "parameters": [
                    {
                        "name": "next_question_id",
                        "in": "path",
                        "description": "Search by next_question_id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "defined_answer_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "question_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/nextQuestion/delete/{next_question_id}": {
            "post": {
                "tags": [
                    "Next Question"
                ],
                "summary": "Delete a Next Question",
                "operationId": "deleteNextQuestion",
                "parameters": [
                    {
                        "name": "next_question_id",
                        "in": "path",
                        "description": "Search by next_question_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/userFriend/all": {
            "get": {
                "tags": [
                    "User Friend"
                ],
                "summary": "Get all the friends from the logged user",
                "description": "Returns a list of friends",
                "operationId": "getNextQuestionsFromSurvey",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "definedAnswer": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_friend_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "user_id_from": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "user_id_to": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "date_related": {
                                                        "type": "string",
                                                        "example": "17/05/2022"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}